+++ lib/blueprint/model.dart
@@ -1,6 +1,4 @@
 part of '../blueprint.dart';
 
 /// Quantum of Data / API / Task interaction — CRUD & low-level functionality only.
-abstract base class Repository {
-  
-}
+abstract base class Model {}